Mouseover color test
colortest.py lets people see an accurate representation of the sheet in the main program, and play around with the background mouseover color which indicates where the user will press. Uploaded for feedback purposes.

The program purposefully lacks complete control of the displayable colors. The user chooses an RGB or CMY base color and plays around with how light/dark it is. I might upload a different version with more free control later.